src/test/java/org/example/springboard/controller/PostControllerTest.java
package org.example.springboard.controller;

import org.example.springboard.service.PostService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(PostController.class)
class PostControllerTest {

    @Autowired
    MockMvc mockMvc;

    @MockBean
    PostService postService;

    @Test
    void 게시글_작성_테스트() throws Exception{

        mockMvc.perform(post("/write")
                .param("title", "제목")
                .param("content", "내용")
        ).andExpect(status().is3xxRedirection());

        verify(postService).addPost(any());
    }
}